fix bug

+++ viphxin/xingo/cluster/rpcprotocol.go
@@ -28,7 +28,7 @@ func (this *RpcServerProtocol) GetDataPack() iface.Idatapack {
 	return this.rpcDatapack
 }
 
-func (this *RpcServerProtocol) ManualMsgPush(msgId uint32, data []byte, pid uint32, fconn iface.Iconnection) {
+func (this *RpcServerProtocol) ManualMsgPush(msgId uint32, data []byte, pid uint64, fconn iface.Iconnection) {
 	logger.Info("invalid ManualMsgPush rpcprotocol, fatal error !")
 }
